Se hacen correciones a los endpoints de listar y generar bill

+++ src/main/java/co/edu/ufps/transaccion/models/responses/BillResponse.java
@@ -7,7 +7,7 @@ public class BillResponse {
 
     private Integer type;
 
-    private Integer value;
+    private Float value;
 
     private String observation;
 }
